refactor: move the initial items into their own packages
