+++ app/src/main/java/com/chesire/zwei/view/onboarding/initial/welcome/WelcomeFragment.kt
@@ -1,4 +1,4 @@
-package com.chesire.zwei.view.onboarding.initial
+package com.chesire.zwei.view.onboarding.initial.welcome
 
 import android.content.Context
 import android.os.Bundle
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.chesire.zwei.R
+import com.chesire.zwei.view.onboarding.initial.InitialInteractor
 import kotlinx.android.synthetic.main.fragment_welcome.buttonNext
 
 class WelcomeFragment : Fragment() {
