+++ app/src/main/res/layout/fragment_request.xml
@@ -1,8 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
     <TextView
         android:id="@+id/textRequest"
         android:layout_width="wrap_content"
@@ -15,4 +18,4 @@
         android:layout_height="wrap_content"
         android:text="Next" />
 
-</LinearLayout>
+</ScrollView>
